Rename scope.js to map.js
