+++ src/expression/node/Node.js
@@ -2,9 +2,9 @@ import { isNode } from '../../utils/is.js'
 
 import { keywords } from '../keywords.js'
 import { deepStrictEqual } from '../../utils/object.js'
-import { hasScopeProperty } from '../../utils/scope.js'
+import { hasScopeProperty } from '../../utils/map.js'
 import { factory } from '../../utils/factory.js'
-import { createScope } from '../../utils/scope.js'
+import { createScope } from '../../utils/map.js'
 
 const name = 'Node'
 const dependencies = ['mathWithTransform']
