+++ test/unit-tests/expression/node/FunctionNode.test.js
@@ -2,7 +2,7 @@
 import assert from 'assert'
 
 import math from '../../../../src/defaultInstance.js'
-import { toObject } from '../../../../src/utils/scope.js'
+import { toObject } from '../../../../src/utils/map.js'
 const Node = math.Node
 const ConstantNode = math.ConstantNode
 const SymbolNode = math.SymbolNode
